// Client/src/app/core/modules/openapi/model/completion-service-request-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CompletionServiceRequestDTO { 
    bulletPoints: string;
    contextPrompt?: string;
    titleEng?: string;
    levelEng?: string;
    languageEng?: CompletionServiceRequestDTO.LanguageEngEnum;
    frequencyEng?: string;
    credits?: number;
    durationEng?: string;
    hoursTotal?: number;
    hoursSelfStudy?: number;
    hoursPresence?: number;
    examinationAchievementsEng?: string;
    repetitionEng?: string;
    recommendedPrerequisitesEng?: string;
    contentEng?: string;
    learningOutcomesEng?: string;
    teachingMethodsEng?: string;
    mediaEng?: string;
    literatureEng?: string;
    responsiblesEng?: string;
    lvSwsLecturerEng?: string;
}
export namespace CompletionServiceRequestDTO {
    export type LanguageEngEnum = 'English' | 'German';
    export const LanguageEngEnum = {
        English: 'English' as LanguageEngEnum,
        German: 'German' as LanguageEngEnum
    };
}